+++ helm/flink-operator/crds/flinksessionjobs.flink.apache.org-v1.yml
@@ -79,14 +79,6 @@ spec:
                         type: integer
                     type: object
                 type: object
-              jobManagerDeploymentStatus:
-                enum:
-                - READY
-                - DEPLOYED_NOT_READY
-                - DEPLOYING
-                - MISSING
-                - ERROR
-                type: string
               reconciliationStatus:
                 properties:
                   success:
